delegated arguments to cli
